Reconciling/upgrading the random number generator.
Using a 32-bit implementation of an extended-precision random number generator across the board in both the C and Fortran versions.

+++ C2021/extern.h
@@ -153,7 +153,7 @@ bool rappl1(int);
 
 // rtim.c:
 void intime(int *, int *, int *);
-void inirnd(int);
+void inirnd(long);
 int rnd(int);
 void exit_(void);
